add SignInPage, refactor styles in pages, replaced router in folder
