+++ src/pagesMinin/Learn.page.tsx
@@ -1,7 +1,9 @@
-import { useParams } from 'react-router-dom'
+import { Link, useParams } from 'react-router-dom'
 
+import ArrowBackOutline from '@/assets/icons/svg/ArrowBackOutline'
 import Typography from '@/components/ui/Typography/Typography'
 import { Card } from '@/components/ui/card'
+import { Page } from '@/pagesMinin/componentsMinin/Page/Page'
 
 import s from './learn.page.module.scss'
 
@@ -12,10 +14,19 @@ export const LearnPage = () => {
   const { data: deck } = useGetDeckByIdQuery({ id: deckId ?? '' })
 
   return (
-    <Card className={s.card}>
-      <Typography className={s.heading} variant={'h1'}>
-        Learn {deck?.name}
-      </Typography>
-    </Card>
+    <Page mt={'24px'}>
+      {/*<h1>Текст для вашей рекламы</h1>*/}
+      <div className={s.headingFirstRow}>
+        <Typography as={Link} style={{ textDecoration: 'none' }} to={'/'} variant={'body2'}>
+          <ArrowBackOutline className={s.backIcon} />
+          Back to Deck List
+        </Typography>
+      </div>
+      <Card className={s.card}>
+        <Typography className={s.centerText} variant={'h1'}>
+          Learn {deck?.name}
+        </Typography>
+      </Card>
+    </Page>
   )
 }
