+++ src/pagesMinin/SignIn.page.tsx
@@ -1,0 +1,10 @@
+import { SignIn } from '@/components/auth/sign-in/sign-in'
+import { Page } from '@/pagesMinin/componentsMinin/Page/Page'
+
+export const SignInPage = () => {
+  return (
+    <Page mt={'36px'}>
+      <SignIn />
+    </Page>
+  )
+}
