+++ src/pagesMinin/cardsPage.module.scss
@@ -67,3 +67,7 @@
 .learnCards {
   text-decoration: none;
 }
+
+.backIcon {
+  margin-right: 6px;
+}
